feat : 데일리 페이지 분할
