+++ src/app/daily/ranking/page.tsx
@@ -1,0 +1,9 @@
+import { Ranking } from "@/src/components/Ranking/ranking";
+
+export default function RankingPage() {
+  return (
+    <div>
+      <Ranking />
+    </div>
+  );
+}
